Master

+++ Backend/FoodiFavs/Controllers/ReviewController.cs
@@ -143,7 +143,7 @@ namespace FoodiFavs.Controllers
 
                 //To count reviews number for each user
                 user.ReviewCount++;
-                user.TotalPoints+=5;
+                user.TotalPoints+=50;
                 restaurant.ReviewCount++;
 
                 await _db.SaveChangesAsync();
@@ -199,7 +199,7 @@ namespace FoodiFavs.Controllers
             else
             {
                 // If points already exist, update the points
-                userRestaurantPoints.PointsForEachRestaurant += 5;
+                userRestaurantPoints.PointsForEachRestaurant += 50;
                 
                 notification.Message = $"{user.UserName} You've earned 5 points for your contribution y!";
                 notification.ReviewId=model.Id;
@@ -219,7 +219,8 @@ namespace FoodiFavs.Controllers
                 }
             }
             userRestaurantPoints.AllPoints=user.TotalPoints;
-            
+            user.UnReadNotiNum = user.UnReadNotiNum ?? 0;
+            user.UnReadNotiNum++;
             _db.Notifications.Add(notification);
             await _db.SaveChangesAsync();
  
@@ -235,23 +236,25 @@ namespace FoodiFavs.Controllers
             //Find all the Users whose following the blogger 
             var followers = _db.FavoriteBloggers
                .Where(f => f.BloggerId == user.Id) //Make suer that we search in the same blogger
-               .Select(f => f.UserId) //Get the followers Id
+               .Select(f => f.User) //Get the followers Id
                .ToList();
 
             //Loob to notify to all Followers
-            foreach (var followerId in followers)
+            foreach (var follower in followers)
             {
                 var notificationReview = new Notification
                 {
-                    UserId = followerId, 
-                    Message = $"{user.UserName}, your favorite blogger, has written a new review for {restaurant.Name}!",
+                    UserId = follower.Id, 
+                    Message = $"your favorite blogger {user.UserName} , has written a new review for {restaurant.Name}!",
                     CreatedAt = DateTime.Now,
                     IsRead = false,
-                    //BloggertId= followerId,
+                    BloggertId= follower.Id,
                     ReviewId=model.Id,
                     RestaurantId=restaurant.Id,
-                    NotificationType="Review"
+                    NotificationType="Review",
                 };
+                follower.UnReadNotiNum = follower.UnReadNotiNum ?? 0;
+                follower.UnReadNotiNum++;
                 _db.Notifications.Add(notificationReview);
             }
             await _db.SaveChangesAsync();
@@ -281,19 +284,38 @@ namespace FoodiFavs.Controllers
             var TopRate = _db.TopReviewForUsers.FirstOrDefault(u => u.UserId==userId);
             var points = _db.Points.FirstOrDefault(p => p.UserId==userId);
             var restaurants = _db.Restaurants.FirstOrDefault(r => r.Id==Review.RestaurantId);
-            var notification=_db.Notifications.FirstOrDefault(n => n.UserId==userId && n.ReviewId==Review.Id);
+            var notifications = _db.Notifications.Where(n => n.ReviewId == Review.Id).ToList();
             var order=_db.Orders.FirstOrDefault(o => o.ReviewId==Review.Id);
+            var likes = _db.Likes.Where(l => l.ReviewId == Review.Id).ToList();
 
-            _db.Orders.Remove(order);
-            _db.Notifications.Remove(notification);
-            
+
+            if (order!=null) 
+            {
+                _db.Orders.Remove(order);
+            }
+            if (likes.Any())
+            {
+                _db.Likes.RemoveRange(likes);
+            }
+            if (notifications.Any())
+            {
+                _db.Notifications.RemoveRange(notifications); 
+            }
+
             _db.Reviews.Remove(Review);
+  
             _db.SaveChanges();
 
-            
-            restaurants.ReviewCount--;
-            points.PointsForEachRestaurant-=5;
-            points.AllPoints-=5;
+            if (restaurants!=null) 
+            {
+                restaurants.ReviewCount--;
+            }
+            if (points!=null) 
+            {
+                points.PointsForEachRestaurant-=5;
+                points.AllPoints-=5;
+            }
+           
             user.TotalLikes-=Review.Likes;
             user.TotalPoints-=5;
             Review.UserNav.ReviewCount--;
@@ -322,7 +344,7 @@ namespace FoodiFavs.Controllers
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [HttpPut("UpadteReview-ById")]
-        [Authorize(Roles ="Admin")]
+        //[Authorize(Roles ="Admin")]
         public IActionResult UpdateReview(int Id, [FromBody] ReviewDto obj)
         {
 
@@ -347,9 +369,18 @@ namespace FoodiFavs.Controllers
                 return Unauthorized(); // Return 401 if user is not logged in
             }
             var user = _db.Users.FirstOrDefault(u => u.UserName == userId);
-            var existingLike = _db.Likes.FirstOrDefault((l => l.UserId == user.Id && l.ReviewId == ReviewId));
+         
             var Review = _db.Reviews.FirstOrDefault(r => r.Id == ReviewId);
+            if (Review == null) 
+            {
+                return BadRequest("No Review with this Id");
+            }
+            var existingLike = _db.Likes.FirstOrDefault((l => l.ReviewId == ReviewId && l.UserId==user.Id));
             var reviewer = _db.Users.FirstOrDefault(u => u.Id == Review.UserId);
+            if (reviewer == null) 
+            {
+                return BadRequest("Sign Up First");
+            }
             if (existingLike == null)
             {
                 var Like = new Like
@@ -361,7 +392,6 @@ namespace FoodiFavs.Controllers
 
                 Review.Likes++;
 
-               
                 if (reviewer != null)
                 {
                     reviewer.TotalLikes++;
@@ -377,8 +407,11 @@ namespace FoodiFavs.Controllers
                         CreatedAt = DateTime.UtcNow,
                         IsRead = false,
                         ReviewId=ReviewId,
-                        NotificationType="Like"
+                        NotificationType="Like",
+                        RestaurantId=Review.RestaurantId
+                        
                     };
+                    user.UnReadNotiNum++;
                     _db.Notifications.Add(notification);
                 }
 
@@ -387,8 +420,14 @@ namespace FoodiFavs.Controllers
             {
                 _db.Likes.Remove(existingLike);
                 Review.Likes--;
-                reviewer.TotalLikes--;
-                user.TotalLikes--;
+                if (reviewer.UserName==user.UserName)
+                {
+                    user.TotalLikes--;
+                }
+                else
+                {
+                    reviewer.TotalLikes--;
+                }
                 var notification = _db.Notifications
                 .FirstOrDefault(n => n.UserId == Review.UserId
                  && n.Message == $"{user.UserName} liked your review.");
